+++ src/app/layout.tsx
@@ -4,6 +4,7 @@ import type { Metadata } from "next";
 import { Inter, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { Toaster } from "sonner";
+import Script from "next/script";
 
 const inter = Inter({ variable: "--font-inter", subsets: ["latin"] });
 const geistMono = Geist_Mono({ variable: "--font-geist-mono", subsets: ["latin"] });
@@ -46,6 +47,23 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       </head>
       <body className="font-sans antialiased h-full">
         <Toaster position="top-right" />
+        <Script id="tracelead-chatwoot" strategy="afterInteractive">
+          {`(function(d,t){
+  var allowed=["/bni","/contact","/pricing","/saiba-mais","/sob-demanda","/sobre-nos","/terms","/privacy","/signup","/login"];
+  if(location.pathname!=="/"&&!allowed.some(p=>location.pathname.startsWith(p))) return;
+  var BASE_URL="https://platform.tracelead.com.br";
+  var g=d.createElement(t),s=d.getElementsByTagName(t)[0];
+  g.src=BASE_URL+"/packs/js/sdk.js";
+  g.defer=true;g.async=true;
+  s.parentNode.insertBefore(g,s);
+  g.onload=function(){
+    window.chatwootSDK.run({
+      websiteToken:"EioqiGzk1toeBkQgLiRNxMuG",
+      baseUrl:BASE_URL
+    });
+  };
+})(document,"script");`}
+        </Script>
         {children}
       </body>
     </html>
